refactor(sync): Extract shared logic into SyncTaskProcessorBase
- Create abstract base class `SyncTaskProcessorBase` for sync processors
- Move all shared fields, batch collections, and methods to base class
- Full sync and delta sync processors now inherit from base class
- Unify obsolete CSO deletion: full sync now uses batching like delta sync
  (reduces database round trips for obsolete CSO deletions)
- Update test to use RemoveRange mock for batch deletions

Closes #252
